+++ src/providers/AliasCompletionProvider.js
@@ -1,0 +1,19 @@
+const { traverseForAlias } = require('./AliasDefinitionProvider');
+
+class AliasCompletionProvider {
+  provideCompletionItems(document) {
+    // look for aliases
+    const aliases = traverseForAlias(document.fileName);
+    // prepare completions list
+    const completions = aliases.map(a => ({
+      label: a.name,
+      // type of completion is Variable
+      kind: 11
+    }));
+    return {
+      items: completions
+    };
+  }
+}
+
+module.exports = AliasCompletionProvider;